requirements.txt
flask
python-dotenv
anthropic
google-genai
openai